modify gitignore, move utility files to utility folder
