--- DiscordBot/commands.js
@@ -1,13 +1,0 @@
-module.exports = {
-  attack: require('./commands/mud/attack'),
-  describe: require('./commands/mud/describe'),
-  item: require('./commands/mud/item'),
-  help: require('./commands/mud/help'),
-  inventory: require('./commands/mud/inventory'),
-  look: require('./commands/mud/look'),
-  move: require('./commands/mud/move'),
-  start: require('./commands/mud/start'),
-  stats: require('./commands/mud/stats'),
-  sudoku: require('./commands/mud/sudoku'),
-  talk: require('./commands/mud/talk')
-};
