+++ DiscordBot/commands/mud/attack.js
@@ -1,6 +1,6 @@
-const { deleteMessage, bigCheck, respawn, commandPrefix } = require('../../globals.js');
+const { deleteMessage, bigCheck, respawn, commandPrefix } = require('../../utilities/globals');
 const commando = require('discord.js-commando');
-const db = require('../../../dbhandler');
+const db = require('../../utilities/dbhandler');
 const { ITEM_CONSTANT } = require('../../Constants/itemConstant');
 const  { COMMAND_CONSTANT } = require('../../Constants/commandConstant');
